Remove the last cell in the daysInMonth array
This array is used in computations where the total number of days from
the previous months in the year is added or subtracted from a day count.
Since December is never among the previous months, the twelfth cell in
the array is not useful.

Fixes #114: an incompatibility with Paul Stoffregen's Time library
arising from both libraries defining identical arrays of constants.