Update ds3231.ino
If the rtc.adjust is inside the the if (rtc.lostpower()) it will not run unless rtc. rtc.lostpower is not running, but if you just uncomment the rtc.adjust it will not set the time unless power was also lost.

I have added a new section for setting the RTC if you need to,
I have with success implemented this changes to my own clone of the library.