Update RTClib.cpp
The current RTC_DS3231::getTemperature function does not handle negative Celsius temperatures correctly.  This correct behavior can be achieved by using a signed 8-bit integer type for the variable "msb" as this allows the sign to be retained for the subsequent float conversion.   Line 1064 was changed to move the "msb" variable to a new line with the int8_t typing. This was tested on an UNO R3 and DS3231 module using a small, frozen blue-ice block with a foam drink cover to achieve negative Celsius temperatures for the DS3231 module.