Consistent time setting statements across examples.
Combines and extends #133 and #134 to additional examples, with some progress toward #132.

Each updated example sets the RTC to compile time after a power loss, as detected by the function specific to that RTC.
The same time-setting statements appear commented out later on for re-setting time on a running RTC, for whatever reason.

The PCF8523 initialized() state wasn't obvious to me. I found that when an unpowered PCF8523 is powered up it goes into standby mode, with Control_3 register set to a default value 0xE0, and in this state initialized() will return false. So it is closely associated with reporting whether power was lost. I made a brief comment to this effect in the member function.
The behavior is explained in section 8.5.1 of the datasheet https://www.nxp.com/docs/en/data-sheet/PCF8523.pdf

These changes tested on Feather M4 Express and Arduino Mega with DS1307 and PCF8523.
The DS3231 example has only cosmetic changes for consistency.

Fixes #156