Add new function RTC_PCF8523::lostPower()
This function mirrors the existing RTC_DS3231::lostPower()

The PCF8523 sets an Oscillator Stopped (OS) flag anytime its oscillator has stopped or been interrupted. This flag offers a way to verify clock integrity, and to know if power was lost.

To purposely set the OS flag, remove the battery and disconnect the RTC from external power supply for ~1 second or more.
When power is re-applied via battery or supply, the OS flag will remain set, even as the oscillator starts and stabilizes.

RTC_PCF8523::lostPower() returns true if the flag is set, and false only after it has been un-set, for example with RTC_PCF8523::Adjust()

This new function is somewhat similar to an existing function:

RTC_PCF8523::initialized() returns true when power management has been configured, for example with RTC_PCF8523::Adjust(), and false when power management configuration is found in 'standby mode', which is the default after power loss.

For many purposes they may be substitutes for each other, but note the inverted relationship of their return values.

These changes tested on Feather M4 Express and Arduino Mega with PCF8523-equipped Adalogger FeatherWing.

From the PCF8523 datasheet:
https://www.nxp.com/docs/en/data-sheet/PCF8523.pdf#page=20

8.6.1.1 Oscillator STOP flag
The OS flag is set whenever the oscillator is stopped. The flag remains
set until cleared by using the interface. When the oscillator is not running, then the OS flag
cannot be cleared. This method can be used to monitor the oscillator.
The oscillator may be stopped, for example, by grounding one of the oscillator pins, OSCI
or OSCO. The oscillator is also considered to be stopped during the time between
power-on and stable crystal resonance. This time may be in a range of 200 ms to 2 s,
depending on crystal type, temperature, and supply voltage. At power-on, the OS flag is
always set.

fixes #128