Add countdown timer support for PCF8523, and an example.
With this change, the user may easily take advantage of additional timing capabilities of the PCF8523:

1) Enable a recurring 'Countdown Timer' with a range from microseconds to over 10 days.
The countdown timer has an interrupt enabled on the INT1 pin that is exposed as INT or SQW on Adafruit breakout boards.
At the end of every countdown, the INT pin is pulled low for an (optionally) adjustable period of time.
This allows a user to attach an interrupt that executes a piece of code at regular intervals, for sensor data collection or other uses.

2) Enable a 'Second Timer' (1Hz), requiring no parameters or configuration, for a fixed interval of 1 second.
This timer acts very much like a 1-second Countdown Timer, or the existing square wave timer with 'writeSqwPinMode(PCF8523_SquareWave1HZ)'.

The implementation of the Countdown Timer uses the PCF8523's 'Timer B' and there are couple reason for this choice:
- Timer B offers the advantage of programmable low pulse widths.
- Timer B can only function as a Countdown Timer, meaning enablement does not block any alternative uses.

Though not implemented at this time, the PCF8523's 'Timer A' may be used as a Watchdog Timer or Countdown Timer, with fixed width low pulses.
Given most MCUs have built in watchdog timer capability, and considering that Timer A and Timer B share the same INT pin exposed on Adafruit breakout boards, the additional value of Timer A appears to be limited. Still, steps were taken to allow future implementation of Timer A.

Enabling the Countdown or Second Timer causes any existing CLKOUT square wave, configured with writeSqwPinMode(), to halt.
Disabling the Countdown or Second Timer narrowly turns off just the chosen timer, preserving state flags for later reading, and preserving other configuration on the same registers.

A big-hammer function is provided to deconfigure all timers which comes in handy during development due to the battery backup on RTCs that prevents regular resets.

The included Arduino example exercises all functionality with and without interrupts, and attempts to educate a user on techniques and subtleties along the way.

Library source additions documented, tested, reviewed using Adafruit Doxyfile. Minor adjustments to existing documentation made for consistency across PCF8523 functions. Library source also formatted with clang-format.

Tested on PCF8523-equipped Adalogger FeatherWing with:
Arduino Mega 2560
Feather M4 Express
Feather nRF52840 Sense