Document a scenario where the OS flag may not be cleared by adjust().
I think this is how a user could get into such a situation:

1) Start with no battery in RTC.
2) Program Arduino to set the time with adjust() immediately on startup and then check lostPower().
3) Power Arduino on with RTC attached.

The OS flag should be set when the RTC receives supply power.
The crystal should turn on and block clearing the OS flag until stable resonance achieved.
If adjust() executes during stabilization, a follow on call to lostPower() should return true.

Leave comment to alert user to this situation in example and in library.