Fix weekly alarms on the DS3231
For the `DS3231_A1_Day' and `DS3231_A2_Day' alarm modes to work, the RTC
has to know the day of the week, but `adjust()' would write 0 to the
relevant register.

Fix these alarm modes by telling the RTC the day of the week in a format
suitable to it: from 1 (Monday) to 7 (Sunday).

Fixes #193.