RV3032C7: Add support for fetching alarm values
This commit introduces 3 new methods to get Alarm values (in line with DS3231):
  DateTime getAlarm();
  RV3032C7AlarmMode getAlarmMode();
  RV3032C7EventType getAlarmEventType();

In addition, disableAlarm() has been enhanced to properly disable alarm. RV3032C7InvalidDate has been defined to check if the alarm is disabled.

Keywords added

Updated the RV3032C7_alarm example to fetch alarm1. Also
improved the formatting of the serial logging and added the possibility to disable the alarm.